refactor for live dialog teaser on front
move timeago, timeahead, timeduration to styleguide
