+++ components/Profile/Comments.js
@@ -1,15 +1,11 @@
 import React from 'react'
 import { CommentTeaser, Interaction } from '@project-r/styleguide'
-import timeago from '../../lib/timeago'
 import withT from '../../lib/withT'
 import CommentLink from '../Discussion/CommentLink'
 
 const Comments = ({ t, comments }) => {
   if (!comments || !comments.totalCount) {
     return null
-  }
-  const timeagoFromNow = createdAtString => {
-    return timeago(t, (new Date() - Date.parse(createdAtString)) / 1000)
   }
   return (
     <div>
@@ -33,7 +29,6 @@ const Comments = ({ t, comments }) => {
             tags={comment.tags}
             parentIds={comment.parentIds}
             discussion={discussion}
-            timeago={timeagoFromNow}
             Link={CommentLink}
             t={t}
           />
